+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [2.1.0] - 2021/10/22
 
-* Saving and setting themes by key.
+* Saving and setting themes by key: `Themed.save()`, `Themed.setThemeByKey()` etc.
+* Fixed https://github.com/marcglasberg/themed/issues/1
 
 ## [2.0.5] - 2021/09/20